Enable warnings for constructor initialization order and for "variable x (#919)
overrides previous variable of the same name" and fix all the results.

+++ src/game/logic/system/cratesystem.cpp
@@ -25,7 +25,7 @@ CrateSystem::CrateSystem()
 
 CrateSystem::~CrateSystem()
 {
-    for (int i = 0; i < m_crateTemplateVector.size(); i++) {
+    for (size_t i = 0; i < m_crateTemplateVector.size(); i++) {
         CrateTemplate *tmplate = m_crateTemplateVector[i];
 
         if (tmplate != nullptr) {
